Rename module auth folder
