+++ api/src/modules/scenarios/scenarios.service.ts
@@ -11,7 +11,7 @@ import { UpdateScenarioDto } from 'modules/scenarios/dto/update.scenario.dto';
 import { ScenarioInterventionsService } from 'modules/scenario-interventions/scenario-interventions.service';
 import { ScenarioIntervention } from 'modules/scenario-interventions/scenario-intervention.entity';
 import { SelectQueryBuilder } from 'typeorm';
-import { ScenariosAccessControl } from 'modules/authorization/modules/scenarios.access-control.service';
+import { ScenariosAccessControl } from 'modules/authorization/formodule/scenarios.access-control.service';
 
 @Injectable()
 export class ScenariosService extends AppBaseService<
